+++ build_vsix_package.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 
 npm install
+npm run esbuild
 npm run vsix