+++ src/extension/extension.js
@@ -12,7 +12,7 @@ const vscode = require('vscode');
 //-----------------------------------------------------------------------------------------------//
 
 global._sourcebase = path.resolve(__dirname, "..");
-const BIND = function(_module) { _module.paths.push(global._sourcebase); };
+const BIND = function(_module) { if (null != _module.paths) { _module.paths.push(global._sourcebase); }};
 global.BIND = BIND
 BIND(module);