+++ src/project/toolkit.js
@@ -26,6 +26,8 @@ class Toolkit {
 
         this.isCpp = (this.isLLVM || this.isCC65 || this.isOscar64);
         this.isAssembler = (this.isAcme || this.isKick);
+
+        this.hasProblemMatcher = (this.isOscar64 || this.isKick);
     }
 
     static fromName(name) {